rename package db to database
